Refactor API router structure by splitting modules
