+++ fia_api/fia_api.py
@@ -18,7 +18,12 @@ from fia_api.exception_handlers import (
     missing_script_handler,
     unsafe_path_handler,
 )
-from fia_api.router import ROUTER
+from fia_api.routers.extras import ExtrasRouter
+from fia_api.routers.instrument_specs import InstrumentSpecRouter
+from fia_api.routers.job_creation import JobCreationRouter
+from fia_api.routers.jobs import JobsRouter
+from fia_api.routers.k8s import k8s_router
+from fia_api.routers.scripts import ScriptRouter
 
 stdout_handler = logging.StreamHandler(stream=sys.stdout)
 logging.basicConfig(
@@ -42,7 +47,12 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(ROUTER)
+app.include_router(ExtrasRouter)
+app.include_router(InstrumentSpecRouter)
+app.include_router(JobCreationRouter)
+app.include_router(JobsRouter)
+app.include_router(k8s_router)
+app.include_router(ScriptRouter)
 
 app.add_exception_handler(MissingRecordError, missing_record_handler)
 app.add_exception_handler(MissingScriptError, missing_script_handler)
